refactor: Move atest CLI/WAV-parsing logic into cmd/samoyed-atest
atest.go isn't a self-contained tool like the ones already moved
(kissutil, tnctest, walk96, ...): it also defines the audio_get/
ptt_set/get_input dispatchers and dlq_rec_frame_fake that core
production code calls unqualified, gated by ATEST_C. Those reference
unexported types (packet_t, fec_type_t) and can't leave package
direwolf, and the config-building/decode-loop code depends on dozens
of other unexported symbols (audio_s, MODEM_*, demod_get_sample, ...).

So only what's genuinely self-contained moves to
cmd/samoyed-atest/main.go: CLI flag parsing/validation, WAV header
parsing, and per-file/summary output. The rest (config building,
decode loop, ATEST_C dispatchers/fakes) stays in a new
src/atest_wrapper.go behind three new exported entry points
(AtestConfigure, AtestDecodeWAV, AtestDCDCounts), mirroring the
existing src/cm108_wrapper_linux.go pattern.
