--- src/atest.go
@@ -1,889 +1,0 @@
-/* Test fixture for the Dire Wolf demodulators */
-//nolint:gochecknoglobals
-package direwolf
-
-/*-------------------------------------------------------------------
- *
- * Purpose:     Test fixture for the Dire Wolf demodulators.
- *
- * Inputs:	Takes audio from a .WAV file instead of the audio device.
- *
- * Description:	This can be used to test the demodulators under
- *		controlled and reproducible conditions for tweaking.
- *
- *		For example
- *
- *		(1) Download WA8LMF's TNC Test CD image file from
- *			http://wa8lmf.net/TNCtest/index.htm
- *
- *		(2) Burn a physical CD.
- *
- *		(3) "Rip" the desired tracks with Windows Media Player.
- *			Select .WAV file format.
- *
- *		"Track 2" is used for most tests because that is more
- *		realistic for most people using the speaker output.
- *
- *
- * 	Without ONE_CHAN defined:
- *
- *	  Notice that the number of packets decoded, as reported by
- *	  this test program, will be twice the number expected because
- *	  we are decoding the left and right audio channels separately.
- *
- *
- * 	With ONE_CHAN defined:
- *
- *	  Only process one channel.
- *
- *--------------------------------------------------------------------*/
-
-import (
-	"bufio"
-	"encoding/binary"
-	"errors"
-	"fmt"
-	"io"
-	"math"
-	"os"
-	"strconv"
-	"strings"
-	"time"
-	"unicode"
-	"unsafe"
-
-	"github.com/spf13/pflag"
-)
-
-type atest_header_t struct {
-	RIFF     [4]byte /* "RIFF" */
-	Filesize int32   /* file length - 8 */
-	WAVE     [4]byte /* "WAVE" */
-}
-
-type atest_chunk_t struct {
-	Id       [4]byte /* "LIST" or "fmt " */
-	Datasize int32
-}
-
-type atest_format_t struct {
-	Wformattag      int16 /* 1 for PCM. */
-	Nchannels       int16 /* 1 for mono, 2 for stereo. */
-	Nsamplespersec  int32 /* sampling freq, Hz. */
-	Navgbytespersec int32 /* = nblockalign*nsamplespersec. */
-	Nblockalign     int16 /* = wbitspersample/8 * nchannels. */
-	Wbitspersample  int16 /* 16 or 8. */
-}
-
-type atest_wav_data_t struct {
-	Data     [4]byte /* "data" */
-	Datasize int32
-}
-
-var ATEST_C = false
-
-var header atest_header_t
-var chunk atest_chunk_t
-var format atest_format_t
-var wav_data atest_wav_data_t
-
-var atestFP *os.File
-var atestBuf *bufio.Reader
-var e_o_f bool
-var packets_decoded_one = 0
-
-var my_audio_config *audio_s
-
-var space_gain [MAX_SUBCHANS]float64
-
-var sample_number = -1 /* Sample number from the file. */
-/* Incremented only for channel 0. */
-/* Use to print timestamp, relative to beginning */
-/* of file, when frame was decoded. */
-
-// command line options.
-
-var h_opt = false // Hexadecimal display of received packet.
-var d_o_opt = 0   // "-d o" option for DCD output control. */
-var dcd_count = 0
-var dcd_missing_errors = 0
-
-const EXPERIMENT_G = true
-const EXPERIMENT_H = true
-
-func AtestMain() {
-	ATEST_C = true
-
-	var count [MAX_SUBCHANS]int // Experiments G and H
-
-	TextColorInit(1)
-	text_color_set(DW_COLOR_INFO)
-
-	my_audio_config = new(audio_s)
-
-	/*
-	 * First apply defaults.
-	 */
-
-	my_audio_config.adev[0].num_channels = DEFAULT_NUM_CHANNELS
-	my_audio_config.adev[0].samples_per_sec = DEFAULT_SAMPLES_PER_SEC
-	my_audio_config.adev[0].bits_per_sample = DEFAULT_BITS_PER_SAMPLE
-
-	for channel := range MAX_RADIO_CHANS {
-		my_audio_config.achan[channel].modem_type = MODEM_AFSK
-
-		my_audio_config.achan[channel].mark_freq = DEFAULT_MARK_FREQ
-		my_audio_config.achan[channel].space_freq = DEFAULT_SPACE_FREQ
-		my_audio_config.achan[channel].baud = DEFAULT_BAUD
-
-		my_audio_config.achan[channel].profiles = "A"
-
-		my_audio_config.achan[channel].num_freq = 1
-		my_audio_config.achan[channel].offset = 0
-
-		my_audio_config.achan[channel].fix_bits = RETRY_NONE
-
-		my_audio_config.achan[channel].sanity_test = SANITY_APRS
-		// my_audio_config.achan[channel].sanity_test = SANITY_AX25;
-		// my_audio_config.achan[channel].sanity_test = SANITY_NONE;
-	}
-
-	var bitrateStr = pflag.StringP("bitrate", "B", strconv.Itoa(DEFAULT_BAUD), `Bits/second for data.  Proper modem automatically selected for speed.
-300 bps defaults to AFSK tones of 1600 & 1800.
-1200 bps uses AFSK tones of 1200 & 2200.
-2400 bps uses QPSK based on V.26 standard.
-4800 bps uses 8PSK based on V.27 standard.
-9600 bps and up uses K9NG/G3RUH standard.
-AIS for ship Automatic Identification System.
-EAS for Emergency Alert System (EAS) Specific Area Message Encoding (SAME).`)
-	var g3ruh = pflag.BoolP("g3ruh", "g", false, "Use G3RUH modem rather than default for data rate.")
-	var bpsk = pflag.BoolP("bpsk", "k", false, "Use BPSK modem rather than default for data rate.")
-	var direwolf15compat = pflag.BoolP("direwolf-15-compat", "j", false, "2400 bps QPSK compatible with direwolf <= 1.5.")
-	var mfj2400compat = pflag.BoolP("mfj-2400-compat", "J", false, "2400 bps QPSK compatible with MFJ-2400.")
-	var modemProfile = pflag.StringP("modem-profile", "P", "", "Select the demodulator type such as D (default for 300 bps), E+ (default for 1200 bps), PQRS for 2400 bps, etc.")
-	var decimate = pflag.IntP("decimate", "D", 0, "Divide audio sample rate by n. 0 is auto-select.")
-	var upsample = pflag.IntP("upsample", "U", 0, "Upsample for G3RUH to improve performance when the sample rate to baud ratio is low.")
-	var fixBits = pflag.IntP("fix-bits", "F", 0, `Amount of effort to try fixing frames with an invalid CRC.
-0 (default) = consider only correct frames.
-1 = Try to fix only a sigle bit.
-Higher values = Try modifying more bits to get a good CRC.`)
-	var errorIfLessThan = pflag.IntP("error-if-less-than", "L", -1, "Error if less than this number decoded.")
-	var errorIfGreaterThan = pflag.IntP("error-if-greater-than", "G", -1, "Error if greater than this number decoded.")
-	var channel0 = pflag.BoolP("channel-0", "0", false, "Use channel 0 (left) of stereo audio (default).")
-	var channel1 = pflag.BoolP("channel-1", "1", false, "Use channel 1 (right) of stereo audio.")
-	var channel2 = pflag.BoolP("channel-2", "2", false, "Use both channels of stereo audio.")
-	var hexDisplay = pflag.BoolP("hex-display", "h", false, "Print frame contents as hexadecimal bytes.")
-	var bitErrorRate = pflag.Float64P("bit-error-rate", "e", 0.0, "Receive Bit Error Rate (BER).")
-	var debugFlags = pflag.StringSliceP("debug", "d", []string{}, `Debug (repeat for increased verbosity).
-x = FX.25
-o = DCD output control
-2 = IL2P`)
-	var help = pflag.Bool("help", false, "Display help text.")
-
-	pflag.Usage = func() {
-		fmt.Fprintf(os.Stderr, "%s is a test application which decodes AX.25 frames from audio recordings.\n", os.Args[0])
-		fmt.Fprintf(os.Stderr, "This provides an easy way to test decoding performance and functionality much quicker than normal real-time.\n")
-		fmt.Fprintf(os.Stderr, "\n")
-		fmt.Fprintf(os.Stderr, "Usage: %s [OPTION]... <WAV FILE>...\n", os.Args[0])
-		pflag.PrintDefaults()
-		fmt.Fprintf(os.Stderr, "\n")
-		fmt.Fprintf(os.Stderr, "Examples:\n")
-		fmt.Fprintf(os.Stderr, "\n")
-		fmt.Fprintf(os.Stderr, "$ gen_packets -o test1.wav\n")
-		fmt.Fprintf(os.Stderr, "$ atest test1.wav\n")
-		fmt.Fprintf(os.Stderr, "\n")
-		fmt.Fprintf(os.Stderr, "$ gen_packets -B 300 -o test3.wav\n")
-		fmt.Fprintf(os.Stderr, "$ atest -B 300 test3.wav\n")
-		fmt.Fprintf(os.Stderr, "\n")
-		fmt.Fprintf(os.Stderr, "$ gen_packets -B 9600 -o test9.wav\n")
-		fmt.Fprintf(os.Stderr, "$ atest -B 9600 test9.wav\n")
-		fmt.Fprintf(os.Stderr, "\n")
-		fmt.Fprintf(os.Stderr, "Try different combinations of options to compare decoding performance.\n")
-	}
-
-	// !!! PARSE !!!
-	pflag.Parse()
-
-	if *help {
-		pflag.Usage()
-		os.Exit(1)
-	}
-
-	var d_x_opt = 0
-	var d_2_opt = 0
-
-	for _, debugFlag := range *debugFlags {
-		switch debugFlag {
-		case "x":
-			d_x_opt++
-		case "o":
-			d_o_opt++
-		case "2":
-			d_2_opt++
-		default:
-			fmt.Fprintf(os.Stderr, "Unrecognised debug flag: %s\n", debugFlag)
-			pflag.Usage()
-			os.Exit(1)
-		}
-	}
-
-	if *decimate < 0 || *decimate > 8 {
-		fmt.Fprintf(os.Stderr, "Decimate should be between 0 and 8 inclusive, not %d.\n", *decimate)
-		pflag.Usage()
-		os.Exit(1)
-	}
-
-	my_audio_config.achan[0].decimate = *decimate
-
-	if *upsample != 0 {
-		if *upsample < 1 || *upsample > 8 {
-			fmt.Fprintf(os.Stderr, "Upsample should be between 1 and 4 inclusive, not %d.\n", *upsample)
-			pflag.Usage()
-			os.Exit(1)
-		}
-
-		my_audio_config.achan[0].upsample = *upsample
-	}
-
-	if BitFixLevel(*fixBits) < RETRY_NONE || BitFixLevel(*fixBits) > RETRY_MAX {
-		fmt.Fprintf(os.Stderr, "Fix Bits should be between %d and %d inclusive, not %d.\n", RETRY_NONE, RETRY_MAX, *fixBits)
-		pflag.Usage()
-		os.Exit(1)
-	}
-
-	my_audio_config.achan[0].fix_bits = BitFixLevel(*fixBits)
-
-	var channelFlagCount int
-
-	for _, b := range []bool{*channel0, *channel1, *channel2} {
-		if b {
-			channelFlagCount++
-		}
-	}
-
-	if channelFlagCount == 0 {
-		*channel0 = true
-	}
-
-	if channelFlagCount > 1 {
-		fmt.Fprintf(os.Stderr, "Exactly one of left/right/both channels must be selected.\n")
-		pflag.Usage()
-		os.Exit(1)
-	}
-
-	var decode_only = 0 /* Set to 0 or 1 to decode only one channel.  2 for both.  */
-	if *channel0 {
-		decode_only = 0
-	}
-
-	if *channel1 {
-		decode_only = 1
-	}
-
-	if *channel2 {
-		decode_only = 2
-	}
-
-	my_audio_config.recv_ber = *bitErrorRate
-
-	// Options from atest.c
-	if *hexDisplay {
-		h_opt = true
-	}
-
-	// Hacks for the magic strings
-	var bitrate, bitrateParseErr = strconv.Atoi(*bitrateStr)
-	if *bitrateStr == "AIS" {
-		bitrate = 0xA15A15
-	} else if *bitrateStr == "EAS" {
-		bitrate = 0xEA5EA5
-	} else if bitrateParseErr != nil {
-		fmt.Fprintf(os.Stderr, "Invalid bitrate (should be an integer or 'AIS' or 'EAS'): %s\n", *bitrateStr)
-		pflag.Usage()
-		os.Exit(1)
-	}
-
-	/*
-	 * Set modem type based on data rate.
-	 * (Could be overridden by -g, -j, or -J later.)
-	 */
-	/*    300 implies 1600/1800 AFSK. */
-	/*    1200 implies 1200/2200 AFSK. */
-	/*    2400 implies V.26 QPSK. */
-	/*    4800 implies V.27 8PSK. */
-	/*    9600 implies G3RUH baseband scrambled. */
-
-	my_audio_config.achan[0].baud = bitrate
-
-	/* We have similar logic in direwolf.c, config.c, gen_packets.c, and atest.c, */
-	/* that need to be kept in sync.  Maybe it could be a common function someday. */
-
-	if my_audio_config.achan[0].baud == 100 { // What was this for?
-		my_audio_config.achan[0].modem_type = MODEM_AFSK
-		my_audio_config.achan[0].mark_freq = 1615
-		my_audio_config.achan[0].space_freq = 1785
-	} else if my_audio_config.achan[0].baud < 600 { // e.g. HF SSB packet
-		my_audio_config.achan[0].modem_type = MODEM_AFSK
-		my_audio_config.achan[0].mark_freq = 1600
-		my_audio_config.achan[0].space_freq = 1800
-		// Previously we had a "D" which was fine tuned for 300 bps.
-		// In v1.7, it's not clear if we should use "B" or just stick with "A".
-	} else if my_audio_config.achan[0].baud < 1800 { // common 1200
-		my_audio_config.achan[0].modem_type = MODEM_AFSK
-		my_audio_config.achan[0].mark_freq = DEFAULT_MARK_FREQ
-		my_audio_config.achan[0].space_freq = DEFAULT_SPACE_FREQ
-	} else if my_audio_config.achan[0].baud < 3600 {
-		my_audio_config.achan[0].modem_type = MODEM_QPSK
-		my_audio_config.achan[0].mark_freq = 0
-		my_audio_config.achan[0].space_freq = 0
-		my_audio_config.achan[0].profiles = ""
-	} else if my_audio_config.achan[0].baud < 7200 {
-		my_audio_config.achan[0].modem_type = MODEM_8PSK
-		my_audio_config.achan[0].mark_freq = 0
-		my_audio_config.achan[0].space_freq = 0
-		my_audio_config.achan[0].profiles = ""
-	} else if my_audio_config.achan[0].baud == 0xA15A15 { // Hack for different use of 9600
-		my_audio_config.achan[0].modem_type = MODEM_AIS
-		my_audio_config.achan[0].baud = 9600
-		my_audio_config.achan[0].mark_freq = 0
-		my_audio_config.achan[0].space_freq = 0
-		my_audio_config.achan[0].profiles = " " // avoid getting default later.
-	} else if my_audio_config.achan[0].baud == 0xEA5EA5 {
-		my_audio_config.achan[0].modem_type = MODEM_EAS
-		my_audio_config.achan[0].baud = 521 // Actually 520.83 but we have an integer field here.
-		// Will make more precise in afsk demod init.
-		my_audio_config.achan[0].mark_freq = 2083  // Actually 2083.3 - logic 1.
-		my_audio_config.achan[0].space_freq = 1563 // Actually 1562.5 - logic 0.
-		my_audio_config.achan[0].profiles = "A"
-	} else {
-		my_audio_config.achan[0].modem_type = MODEM_SCRAMBLE
-		my_audio_config.achan[0].mark_freq = 0
-		my_audio_config.achan[0].space_freq = 0
-		my_audio_config.achan[0].profiles = " " // avoid getting default later.
-	}
-
-	if my_audio_config.achan[0].baud < MIN_BAUD || my_audio_config.achan[0].baud > MAX_BAUD {
-		text_color_set(DW_COLOR_ERROR)
-		fmt.Printf("Use a more reasonable bit rate in range of %d - %d.\n", MIN_BAUD, MAX_BAUD)
-		os.Exit(1)
-	}
-
-	/*
-	 * -g option means force g3RUH regardless of speed.
-	 */
-
-	if *g3ruh {
-		my_audio_config.achan[0].modem_type = MODEM_SCRAMBLE
-		my_audio_config.achan[0].mark_freq = 0
-		my_audio_config.achan[0].space_freq = 0
-		my_audio_config.achan[0].profiles = " " // avoid getting default later.
-	}
-
-	if *bpsk {
-		my_audio_config.achan[0].modem_type = MODEM_BPSK
-		my_audio_config.achan[0].mark_freq = 0
-		my_audio_config.achan[0].space_freq = 0
-		my_audio_config.achan[0].profiles = ""
-	}
-
-	/*
-	 * We have two different incompatible flavors of V.26.
-	 */
-	if *direwolf15compat {
-		// V.26 compatible with earlier versions of direwolf.
-		//   Example:   -B 2400 -j    or simply   -j
-		my_audio_config.achan[0].v26_alternative = V26_A
-		my_audio_config.achan[0].modem_type = MODEM_QPSK
-		my_audio_config.achan[0].mark_freq = 0
-		my_audio_config.achan[0].space_freq = 0
-		my_audio_config.achan[0].baud = 2400
-		my_audio_config.achan[0].profiles = ""
-	}
-
-	if *mfj2400compat {
-		// V.26 compatible with MFJ and maybe others.
-		//   Example:   -B 2400 -J     or simply   -J
-		my_audio_config.achan[0].v26_alternative = V26_B
-		my_audio_config.achan[0].modem_type = MODEM_QPSK
-		my_audio_config.achan[0].mark_freq = 0
-		my_audio_config.achan[0].space_freq = 0
-		my_audio_config.achan[0].baud = 2400
-		my_audio_config.achan[0].profiles = ""
-	}
-
-	// Needs to be after -B, -j, -J.
-	if *modemProfile != "" {
-		fmt.Printf("Demodulator profile set to \"%s\"\n", *modemProfile)
-		my_audio_config.achan[0].profiles = *modemProfile
-	}
-
-	my_audio_config.achan[1] = my_audio_config.achan[0]
-
-	if len(pflag.Args()) == 0 {
-		text_color_set(DW_COLOR_ERROR)
-		fmt.Printf("Specify .WAV file name on command line.\n\n")
-		pflag.Usage()
-		os.Exit(1)
-	}
-
-	FX25Init(d_x_opt)
-	il2p_init(d_2_opt)
-
-	var start_time = time.Now()
-	var total_filetime float64
-	var packets_decoded_total = 0
-
-	for _, wavFileName := range pflag.Args() {
-		var err error
-
-		atestFP, err = os.Open(wavFileName) //nolint:gosec // File path from CLI is expected for this tool
-		if err != nil {
-			text_color_set(DW_COLOR_ERROR)
-			fmt.Printf("Couldn't open file %s for read: %s\n", wavFileName, err)
-			// perror ("more info?");
-			os.Exit(1)
-		}
-
-		/*
-		 * Read the file header.
-		 * Doesn't handle all possible cases but good enough for our purposes.
-		 */
-
-		err = binary.Read(atestFP, binary.LittleEndian, &header)
-		if err != nil {
-			text_color_set(DW_COLOR_ERROR)
-			fmt.Printf("WAV file error: Could not read file header: %s\n", err)
-			os.Exit(1)
-		}
-
-		if string(header.RIFF[:]) != "RIFF" || string(header.WAVE[:]) != "WAVE" {
-			text_color_set(DW_COLOR_ERROR)
-			fmt.Printf("This is not a .WAV format file.\n")
-			os.Exit(1)
-		}
-
-		err = binary.Read(atestFP, binary.LittleEndian, &chunk)
-		if err != nil {
-			text_color_set(DW_COLOR_ERROR)
-			fmt.Printf("WAV file error: Could not read chunk header: %s\n", err)
-			os.Exit(1)
-		}
-
-		for string(chunk.Id[:]) != "fmt " {
-			if chunk.Datasize < 0 {
-				text_color_set(DW_COLOR_ERROR)
-				fmt.Printf("WAV file error: Invalid chunk datasize %d.\n", chunk.Datasize)
-				os.Exit(1)
-			}
-
-			_, err = atestFP.Seek(int64(chunk.Datasize)+int64(chunk.Datasize%2), io.SeekCurrent)
-			if err != nil {
-				text_color_set(DW_COLOR_ERROR)
-				fmt.Printf("WAV file error: Could not Seek: %s.\n", err)
-				os.Exit(1)
-			}
-
-			err = binary.Read(atestFP, binary.LittleEndian, &chunk)
-			if err != nil {
-				text_color_set(DW_COLOR_ERROR)
-				fmt.Printf("WAV file error: Could not find \"fmt \" chunk.\n")
-				os.Exit(1)
-			}
-		}
-
-		if chunk.Datasize != 16 && chunk.Datasize != 18 {
-			text_color_set(DW_COLOR_ERROR)
-			fmt.Printf("WAV file error: Need fmt chunk datasize of 16 or 18.  Found %d.\n", chunk.Datasize)
-			os.Exit(1)
-		}
-
-		binary.Read(atestFP, binary.LittleEndian, &format)
-
-		// KG If Datasize > sizeof(format), skip until the actual data
-		var formatSize = int32(unsafe.Sizeof(format))
-		if chunk.Datasize > formatSize {
-			var extra = chunk.Datasize - formatSize
-
-			_, err = atestFP.Seek(int64(extra), io.SeekCurrent)
-			if err != nil {
-				text_color_set(DW_COLOR_ERROR)
-				fmt.Printf("WAV file error: Could not Seek: %s.\n", err)
-				os.Exit(1)
-			}
-		}
-
-		err = binary.Read(atestFP, binary.LittleEndian, &wav_data)
-		if err != nil {
-			text_color_set(DW_COLOR_ERROR)
-			fmt.Printf("WAV file error: Could not read data chunk header: %s\n", err)
-			os.Exit(1)
-		}
-
-		for string(wav_data.Data[:]) != "data" {
-			if wav_data.Datasize < 0 {
-				text_color_set(DW_COLOR_ERROR)
-				fmt.Printf("WAV file error: Invalid chunk datasize %d.\n", wav_data.Datasize)
-				os.Exit(1)
-			}
-
-			_, err = atestFP.Seek(int64(wav_data.Datasize)+int64(wav_data.Datasize%2), io.SeekCurrent)
-			if err != nil {
-				text_color_set(DW_COLOR_ERROR)
-				fmt.Printf("WAV file error: Could not Seek: %s.\n", err)
-				os.Exit(1)
-			}
-
-			err = binary.Read(atestFP, binary.LittleEndian, &wav_data)
-			if err != nil {
-				text_color_set(DW_COLOR_ERROR)
-				fmt.Printf("WAV file error: Could not find \"data\" chunk.\n")
-				os.Exit(1)
-			}
-		}
-
-		if format.Wformattag != 1 {
-			text_color_set(DW_COLOR_ERROR)
-			fmt.Printf("Sorry, I only understand audio format 1 (PCM).  This file has %d.\n", format.Wformattag)
-			os.Exit(1)
-		}
-
-		if format.Nchannels != 1 && format.Nchannels != 2 {
-			text_color_set(DW_COLOR_ERROR)
-			fmt.Printf("Sorry, I only understand 1 or 2 channels.  This file has %d.\n", format.Nchannels)
-			os.Exit(1)
-		}
-
-		if format.Wbitspersample != 8 && format.Wbitspersample != 16 {
-			text_color_set(DW_COLOR_ERROR)
-			fmt.Printf("Sorry, I only understand 8 or 16 bits per sample.  This file has %d.\n", format.Wbitspersample)
-			os.Exit(1)
-		}
-
-		my_audio_config.adev[0].samples_per_sec = int(format.Nsamplespersec)
-		my_audio_config.adev[0].bits_per_sample = int(format.Wbitspersample)
-		my_audio_config.adev[0].num_channels = int(format.Nchannels)
-
-		my_audio_config.chan_medium[0] = MEDIUM_RADIO
-		if format.Nchannels == 2 {
-			my_audio_config.chan_medium[1] = MEDIUM_RADIO
-		}
-
-		text_color_set(DW_COLOR_INFO)
-		fmt.Printf("%d samples per second.  %d bits per sample.  %d audio channels.\n",
-			my_audio_config.adev[0].samples_per_sec,
-			my_audio_config.adev[0].bits_per_sample,
-			(my_audio_config.adev[0].num_channels))
-		// nnum_channels is known to be 1 or 2.
-		var one_filetime = float64(wav_data.Datasize) /
-			float64((my_audio_config.adev[0].bits_per_sample/8)*(my_audio_config.adev[0].num_channels)*my_audio_config.adev[0].samples_per_sec)
-		total_filetime += one_filetime
-
-		fmt.Printf("%d audio bytes in file.  Duration = %.1f seconds.\n",
-			wav_data.Datasize,
-			one_filetime)
-		fmt.Printf("Fix Bits level = %d\n", my_audio_config.achan[0].fix_bits)
-
-		/*
-		 * Initialize the AFSK demodulator and HDLC decoder.
-		 * Needs to be done for each file because they could have different sample rates.
-		 */
-		multi_modem_init(my_audio_config)
-
-		packets_decoded_one = 0
-
-		atestBuf = bufio.NewReader(atestFP)
-
-		e_o_f = false
-		for !e_o_f {
-			for c := 0; c < (my_audio_config.adev[0].num_channels); c++ {
-				/* This reads either 1 or 2 bytes depending on */
-				/* bits per sample.  */
-				var audio_sample = demod_get_sample(ACHAN2ADEV(c))
-
-				if audio_sample >= 256*256 {
-					e_o_f = true
-					continue
-				}
-
-				if c == 0 {
-					sample_number++
-				}
-
-				if decode_only == 0 && c != 0 {
-					continue
-				}
-
-				if decode_only == 1 && c != 1 {
-					continue
-				}
-
-				multi_modem_process_sample(c, audio_sample)
-			}
-
-			/* When a complete frame is accumulated, */
-			/* process_rec_frame, below, is called. */
-		}
-
-		text_color_set(DW_COLOR_INFO)
-		fmt.Printf("\n\n")
-
-		if EXPERIMENT_G {
-			for j := range MAX_SUBCHANS {
-				var db = 20.0 * math.Log10(space_gain[j])
-				fmt.Printf("%+.1f dB, %d\n", db, count[j])
-			}
-		}
-
-		if EXPERIMENT_H {
-			for j := range MAX_SUBCHANS {
-				fmt.Printf("%d\n", count[j])
-			}
-		}
-
-		fmt.Printf("%d from %s\n", packets_decoded_one, wavFileName)
-		packets_decoded_total += packets_decoded_one
-
-		atestFP.Close()
-	}
-
-	var elapsed = time.Since(start_time)
-
-	fmt.Printf("%d packets decoded in %.3f seconds.  %.1f x realtime\n", packets_decoded_total, elapsed.Seconds(), total_filetime/float64(elapsed.Seconds()))
-
-	if d_o_opt > 0 {
-		fmt.Printf("DCD count = %d\n", dcd_count)
-		fmt.Printf("DCD missing errors = %d\n", dcd_missing_errors)
-	}
-
-	if *errorIfLessThan != -1 && packets_decoded_total < *errorIfLessThan {
-		text_color_set(DW_COLOR_ERROR)
-		fmt.Printf("\n * * * TEST FAILED: number decoded is less than %d * * * \n", *errorIfLessThan)
-		os.Exit(1)
-	}
-
-	if *errorIfGreaterThan != -1 && packets_decoded_total > *errorIfGreaterThan {
-		text_color_set(DW_COLOR_ERROR)
-		fmt.Printf("\n * * * TEST FAILED: number decoded is greater than %d * * * \n", *errorIfGreaterThan)
-		os.Exit(1)
-	}
-}
-
-/*
- * Simulate sample from the audio device.
- */
-
-func audio_get_fake(_ int) int {
-	if wav_data.Datasize <= 0 {
-		e_o_f = true
-		return (-1)
-	}
-
-	var data, err = atestBuf.ReadByte()
-	wav_data.Datasize--
-
-	if errors.Is(err, io.EOF) {
-		text_color_set(DW_COLOR_ERROR)
-		dw_printf("Unexpected end of file.\n")
-
-		e_o_f = true
-	}
-
-	// TODO KG Better error handling
-
-	return int(data)
-}
-
-func audio_get(a int) int {
-	if ATEST_C {
-		return audio_get_fake(a)
-	} else {
-		return audio_get_real(a)
-	}
-}
-
-/*
- * This is called when we have a good frame.
- */
-
-func dlq_rec_frame_fake(channel int, subchan int, slice int, pp *packet_t, alevel ALevel, fec_type fec_type_t, retries BitFixLevel, spectrum string) {
-	packets_decoded_one++
-
-	if hdlc_rec_data_detect_any(channel) == 0 {
-		dcd_missing_errors++
-	}
-
-	var stemp = AX25FormatAddrs(pp)
-
-	var info = AX25GetInfo(pp)
-
-	/* Print so we can see what is going on. */
-
-	//TODO: quiet option - suppress packet printing, only the count at the end.
-
-	/* Display audio input level. */
-	/* Who are we hearing?   Original station or digipeater? */
-
-	var h int
-	var heard string
-
-	if ax25_get_num_addr(pp) == 0 {
-		/* Not AX.25. No station to display below. */
-		h = -1
-	} else {
-		h = ax25_get_heard(pp)
-		heard = ax25_get_addr_with_ssid(pp, h)
-	}
-
-	text_color_set(DW_COLOR_DEBUG)
-	dw_printf("\n")
-	dw_printf("DECODED[%d] ", packets_decoded_one)
-
-	/* Insert time stamp relative to start of file. */
-
-	var sec = float64(sample_number) / float64(my_audio_config.adev[0].samples_per_sec)
-	var minutes = int(sec / 60.)
-	sec -= float64(minutes * 60)
-
-	dw_printf("%d:%06.3f ", minutes, sec)
-
-	if h != AX25_SOURCE {
-		dw_printf("Digipeater ")
-	}
-
-	var alevel_text = ax25_alevel_to_text(alevel)
-
-	/* As suggested by KJ4ERJ, if we are receiving from */
-	/* WIDEn-0, it is quite likely (but not guaranteed), that */
-	/* we are actually hearing the preceding station in the path. */
-
-	if h >= AX25_REPEATER_2 &&
-		strings.HasPrefix(heard, "WIDE") &&
-		unicode.IsDigit(rune(heard[4])) &&
-		len(heard) == 5 {
-		var probably_really = ax25_get_addr_with_ssid(pp, h-1)
-
-		heard += " (probably " + probably_really + ")"
-	}
-
-	switch fec_type {
-	case fec_type_fx25:
-		dw_printf("%s audio level = %s   FX.25  %s\n", heard, alevel_text, spectrum)
-	case fec_type_il2p:
-		dw_printf("%s audio level = %s   IL2P  %s\n", heard, alevel_text, spectrum)
-	default:
-		//case fec_type_none:
-		if my_audio_config.achan[channel].fix_bits == RETRY_NONE && !my_audio_config.achan[channel].passall {
-			// No fix_bits or passall specified.
-			dw_printf("%s audio level = %s     %s\n", heard, alevel_text, spectrum)
-		} else {
-			Assert(retries >= RETRY_NONE && retries <= RETRY_MAX) // validate array index.
-			dw_printf("%s audio level = %s   [%s]   %s\n", heard, alevel_text, retries.String(), spectrum)
-		}
-	}
-
-	// Display non-APRS packets in a different color.
-
-	// Display channel with subchannel/slice if applicable.
-
-	if ax25_is_aprs(pp) {
-		text_color_set(DW_COLOR_REC)
-	} else {
-		text_color_set(DW_COLOR_DEBUG)
-	}
-
-	if my_audio_config.achan[channel].num_subchan > 1 && my_audio_config.achan[channel].num_slicers == 1 {
-		dw_printf("[%d.%d] ", channel, subchan)
-	} else if my_audio_config.achan[channel].num_subchan == 1 && my_audio_config.achan[channel].num_slicers > 1 {
-		dw_printf("[%d.%d] ", channel, slice)
-	} else if my_audio_config.achan[channel].num_subchan > 1 && my_audio_config.achan[channel].num_slicers > 1 {
-		dw_printf("[%d.%d.%d] ", channel, subchan, slice)
-	} else {
-		dw_printf("[%d] ", channel)
-	}
-
-	dw_printf("%s", stemp) /* stations followed by : */
-	AX25SafePrint(info, false)
-	dw_printf("\n")
-
-	/*
-	 * -h option for hexadecimal display.  (new in 1.6)
-	 */
-
-	if h_opt {
-		text_color_set(DW_COLOR_DEBUG)
-		dw_printf("------\n")
-		ax25_hex_dump(pp)
-		dw_printf("------\n")
-	}
-
-	/*
-		#if 0		// temp experiment
-
-		#include "decode_aprs.h"
-
-			if (ax25_is_aprs(pp)) {
-
-			  decode_aprs_t A;
-
-			  decode_aprs (&A, pp, 0, NULL);
-
-			  // Temp experiment to see how different systems set the RR bits in the source and destination.
-			  // log_rr_bits (&A, pp);
-
-			}
-		#endif
-	*/
-
-	AX25Delete(pp)
-} /* end fake dlq_append */
-
-var dcd_start_seconds [MAX_RADIO_CHANS]float64
-
-func ptt_set_fake(_ int, channel int, ptt_signal int) {
-	// Should only get here for DCD output control.
-	if d_o_opt > 0 {
-		var t = float64(sample_number) / float64(my_audio_config.adev[0].samples_per_sec)
-
-		text_color_set(DW_COLOR_INFO)
-
-		if ptt_signal != 0 {
-			//sec1 = t;
-			//min1 = (int)(sec1 / 60.);
-			//sec1 -= min1 * 60;
-			//dw_printf ("DCD[%d] = ON    %d:%06.3f\n",  channel, min1, sec1);
-			dcd_count++
-			dcd_start_seconds[channel] = t
-		} else {
-			//dw_printf ("DCD[%d] = off   %d:%06.3f   %3.0f\n",  channel, min, sec, (t - dcd_start_seconds[channel]) * 1000.);
-			var sec1 = dcd_start_seconds[channel]
-			var min1 = (int)(sec1 / 60.)
-			sec1 -= float64(min1 * 60)
-
-			var sec2 = t
-			var min2 = (int)(sec2 / 60.)
-			sec2 -= float64(min2 * 60)
-
-			dw_printf("DCD[%d]  %d:%06.3f - %d:%06.3f =  %3.0f\n", channel, min1, sec1, min2, sec2, (t-dcd_start_seconds[channel])*1000.)
-		}
-	}
-}
-
-func ptt_set(ot int, channel int, ptt_signal int) {
-	if ATEST_C {
-		ptt_set_fake(ot, channel, ptt_signal)
-	} else {
-		ptt_set_real(ot, channel, ptt_signal)
-	}
-}
-
-func get_input_fake(it int, channel int) int {
-	return -1
-}
-
-func get_input(it int, channel int) int {
-	if ATEST_C {
-		return get_input_fake(it, channel)
-	} else {
-		return get_input_real(it, channel)
-	}
-}
